Do not defer on subsequent boot
On the first boot we want to defer until the EdgeNodeCerts have been
published to the controller, but on a subsequent boot we need to proceed
and use a checkpointed config. As part of this we make sure we do not
attempt to restart the attestation if we didn't yet try.

Signed-off-by: eriknordmark <erik@zededa.com>