wwan: make sure the APN is set in the correct profile
Previously, wwan microservice would always set APN for the first PDP
profile (index 1), but some LTE service providers use different profile
number for the data service. For example, in case of Verizon the profile
no. 1 should not be modified and is supposed to be always the same
(APN=vzwims), while data service is supposed to be defined inside
the profile number 3.

wwan microservice should therefore obtain the index of the data service
profile and set the APN there. With qmicli, the data profile is referenced
to as "default" profile.

Additionally, it is better to keep PDP type (IP, IPV6, IPV4V6) unchanged
rather than to assume what the right value should be, which could cause
problems with the registration of the device into the network.

Signed-off-by: Milan Lenco <milan@zededa.com>