Support DNS over TCP
In pillar we use customized dialers for DNS resolvers to enforce the
source IP address to use, thus effectively choosing the output network
interface for a given DNS request.
However, currently we expect that DNS servers always respond on the UDP
port, but they may in fact fallback to use TCP (same port number).
This happends for example when the response message size is too large.
In that case we get "mismatched local address type" error.

I observed this with Azure datastore and a private DNS server, so it
make sense to support DNS over TCP at least for Downloader, but for
completeness I also added this for the controller name resolution.

Signed-off-by: Milan Lenco <milan@zededa.com>