Handle containerd upgrade taking a long time
The logic in upgradeconverter can take a long time; one observed case
did not finish in 7 minutes with 10,000 files and 1.5Gbyte to copy from
/persist/containerd to /persist/vault/containerd. That resulted in the
vault not being unlocked in time for the EVE-OS update to succeed. Here
we tell the vault is unlocked before we run the upgrade converter. Other
microservices which will use the vault continue to check for it having
completed.

Signed-off-by: eriknordmark <erik@zededa.com>