README.md: Mac M1 helps and binfmt clarification
Added dependencies not mentioned in the instructioins to install
`qemu`.

Added steps that may be required for a cross build.  Working from
this README, could not complete the cross build.  Eric recalled the
`redit` article, which allowed completion of the cross build (that is,
`amd64` build on the `aarch64` Mac M1.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>