Manually set ServerName in TLS config breaks proxy listening on HTTPS
We manually set ServerName inside tls.Config, which is then used by tls
package to verify the hostname on the returned certificates and as the SNI
value of the ClientHello TLS message.

However, it is better let the higher-level packages like net/http [1] and
gorilla/websocket [2] to set it automatically from the destination URL.
Setting this manually actually breaks certificate verification when
(non-transparent) network proxy listening on HTTPS (not HTTP) is being
used between the device and the destination.
In that case, the first TLS handshake is being done with the proxy and
it is expected that proxy presents its own certificate, with CN set
to its hostname.

With ServerName set to destination hostname already for this first TLS
handshake, proxy cert verification would fail with:
  proxyconnect tcp: x509: certificate is valid for <proxy-hostname>, not <destination-hostname>

Higher-level packages first set ServerName to the proxy hostname,
then for the subsequent TLS handshake they use the destination hostname.
I checked that if either the proxy presents cert with a wrong CN or the
destination server has cert with unexpected CN (or proxy acting as MITM
generates cert with bad CN), the request fails in every case, returning
a cert verification error as expected even if we do not manually set
ServerName.

Additionally, when diag runs tests towards google.com, it disables
server verification in the TLS config and this stays disabled even for
future diag tests against the controller URL, which is not desired.
This commit fixes this issue as well.

[1]: https://github.com/golang/go/blob/release-branch.go1.16/src/net/http/transport.go#L1511-L1513
[2]: https://github.com/gorilla/websocket/blob/v1.5.0/client.go#L340-L342

Signed-off-by: Milan Lenco <milan@zededa.com>