qemu: Add 'cpu-pin' and 'cpumask' options to QEMU.
Add the 'cpumask' option to set the CPU mask for the threads created by QEMU.
The mask affects all the threads belonging to a VM: both VCPU threads and
non-VCPU threads. If the option is not provided, it's considered to be '-1',
which corresponds to all the available CPUs.  The CPU mask in represented in
the form "d[[,-]d]*". E.g. "0-2" or "0-2,5,6".  CPUs start with 0. For example,
the mask "0,3" would mean that only physical CPUs 0 and 3 are available for the
VM.

Add the 'cpu-pin' option to pin any VCPU thread to a specific CPU. If the
option is set, any VCPU thread will be assigned to a CPU provided with the
'cpumask'. If it's not provided, the VPCU threads can migrate from a CPU to a
CPU within the set of CPUs provided by the 'cpumask' option.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>