kernels,storage-init,mkimage-raw-efi: make ZFS built-in
For the sake of the dump-capture kernel stability make ZFS built-in.
Dump-capture kernel is loaded with 'nomodule' kernel parameter (see
future patches), which prevents any module load, thus makes kernel
stable and makes dump-capture kernel is responsible for only one
thing: collecting a memory dump in /persist volume, which requires
block layer and fs to be functional.

According to the current kernel configuration all block devices and
ext4 fs subsystem are already built-in, so making ZFS as built-in
does not contradict with the whole concept.

Also this patch changes storage-init.sh and install, which load
the zfs module. Now modprobe/rmmod calls are not needed and all
modules parameters are passed through the corresponding sysfs module
entries.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>