pillar: build delve using install command
go get doesn't build with the newer version of linuxkit. If you try to
build with DEV=y you would get:

  'go get' is no longer supported outside a module.

But a simple change to install would not work either, because that
would surface another issue which was hidden before:

  go: github.com/go-delve/delve/cmd/dlv@v1.8.3: cannot query module due to -mod=vendor

To fix that, we let unset the GOFLAGS only for building of the
dev. That would disable the earlier set GOFLAGS=-mod=vendor option.

While at it, also bump the delve version to the latest.

Signed-off-by: Yuri Volchkov <yuri@volch.org>