DNSNotAvail may be incorrectly returned when /etc/hosts is used
Send* methods from pillar/zedcloud package will not call DNS resolver if
the controller's IP address is already known and cached in /etc/hosts by
pillar/cmd/nim/controllerdns.go

Previously, if Send* method used cached controller IP but then the
request failed for some other reason, we would incorretly return DNSNotAvail
error because there was no successful hit detected inside the resolver
(as it was not called at all). This is now handled by this patch. In
this scenario a correct error is returned (whatever it may be),
not DNSNotAvail.

Signed-off-by: Milan Lenco <milan@zededa.com>