compose-image: do not use inplace edit in yq container
The yq container assumes that user id is 1000. Which is true for most
of the systems, but not for all of them. The obvious example is a
shared machine, where only the first user will have uid 1000.

As the result the development build (DEV=y) will work only for uid
1000, and for other users yq will silently fail to update the rootfs
yaml configuration.

Instead inplace editing, lets use the stdout redirection. Yq container
still have permissions to read. And the actual write will happen on
the host in case of redirection.

Also add `set -e` to fail build earlier if something is wrong with the
script.

Signed-off-by: Yuri Volchkov <yuri@volch.org>