images,mkimage-raw-efi: add comments explaining the installer onboot order
I spent quite a while understanding the linuxkit trick in the
pkg/mkimage-raw-efi/Dockerfile which creates the 003-installer
container and injects it *before* the storage-init, where the
order is very crucial: first installer, then storage-init
(installer does the installation and reboots, so storage-init
never runs).

This patch adds explicit comments on the order in order to save
time for a future curious reader.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>