With bootstrap config disable lastresort by default
When device is installed using a single-use EVE installer, the intended
network configuration is available from the very first boot, even before
device onboarding. It therefore does not make much sense to still keep
using last-resort network config as a fallback.

In fact, lastresort is often causing issues, especially when network
outage is outside of the EVE control (external network goes down).
Changing network configuration from the intended state to lastresort
not only cannot help in such cases, but it can also have negative
side effects, such as:
 - prolonged outage: device will stay with lastresort if all DPCs
   are failing and it will only retest the intended config once per
   10 minutes
 - confusing logs and events - customer might get a wrong impression
   that outage was caused by EVE when it unwantedly changed applied config
   (customer will blame EVE instead of troubleshooting outside networks)
 - break NIC pass-through - EVE will mark all ethernets for management
   which will prevent them from being assigned (this can be triggered by
   a race between domainmgr deploying app and NIM falling back to
   lastresort at around the same time)

Instead of changing the default value of lastresort all across the
board, the proposal is to at least change it for the case of device
being installed using a single-use EVE installer (which is a new feature
so this is not changing behavior for older workflows).

Signed-off-by: Milan Lenco <milan@zededa.com>