pillar/domainmng: remove versionning from the cloud-init meta-data file
There is a nasty behavior observed on the VM with a cloud-init
configuration: the field config.UUIDandVersion.Version is
incremented by one each time an application (VM) restarts
from the zedcontrol, which causes cloud-init to be triggered
in the guest and repeat the whole init procedure.

This patch fixes the problem by removing the "version" part
from the instance-id field in the meta-data file in order to
keep cloud-init executed only once.

PS. this patch is a good candidate for a stable branch, thus
    "stable" or whatevert mark is needed here.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>