Remove hardwaremodel -f
This hardware figerprint was using github.com/rackn/gohai which has been
archived and which is also using some godmi packages with unspecified
license. The hardware fingerprint is not used any more so it makes sense
to remove these dependencies.

The update of the vendoring seems to require pulling in a new version of
github.com/digitalocean packages, which in turn means a new qmp monitor
version with an additional argument.

Signed-off-by: eriknordmark <erik@zededa.com>