Update edge-containers to support updated oras-go
In order to support updated oras-go which uses io.ReadFull we should
allocate the buffer for ImageConfig. So let's update edge-containers.
Using ioutil.ReadAll we cannot read from updated version of oras-go as
we will give small buffer to internal io.ReadFull function and will not
be able to return an image from Config function.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>