Implementation of measure-config container
The container is put into onboot section and runs every time the device is
rebooted

CONFIG partition is copied into tmpfs FS by storage-init. Its content is
measured file by file. There are two special lists of files:

1. Files unique for each device e.g. tmp_credential. We do not measure
the content of such files but record the fact that the file exists/
doesn't exist
2. Files that may appear of config partition later e.g. GlobalConfig/override.json
if the file exists its content is measured, if it doesn't we record this fact

NOTE: Device certificates are generated after this container exists but
it happens during the installation and it seems the device is always
rebooted after the installation so the files (file paths to be precise)
will be measured on the next boot. It shouldn't affect real world use cases,
however live image is onboarded will not pass the attestation after reboot

Signed-off-by: Mikhail Malyshev <mikem@zededa.com>