pkg/dom0-ztools: Fix definition of the CPU sets for system services.
The 'dom0_max_vcpus', 'eve_max_vcpus', 'ctrd_max_vcpus' kernel arguments
represent the amounts of CPUs that are assigned to the corresponding CPU sets
when cgroups are created. The CPU set is defined with CPU indexes, that start
with 0, not 1. So, the proper CPU set shoudl look like '0-$((total_amount -
1))'.
Also check the values read from the kernel args. If a value is 0 - set it to a
default one.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>