pkg/pillar: Add a channel for notifications about CPU set changes.
The channel is used by a domain to get messages, when its CPU set may be
changed. The CPU set of the domain with no CPU pinning enabled may be changed
when a domain with CPU pinning enabled is activated or deactivated. In this
case, the domain checks the list of free CPUs and adapts its CPU affinity list.

Also add a function to send the notifications to all domains.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>