Refactoring of dependency checking during packages build
When we want to build one package (for example pillar), we run
`make pkg/pillar` and our expectation is that we will build pillar. But
instead we build all dependencies as well, which is not required
since we decided to not use templates and to embed dependencies
versions directly into Dockerfile of package.
Let's move dependency checking  into `make pkgs` target where we want
to build all packages automatically. This way we will speed up the
build process (dramatically for the first run, noticeable for
subsequent runs).
If we really need changes from another packages to build we should
build the dependency explicitly (for example, using
`make pkg/dom0-tools pkg/pillar`).

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>