pkg/pillar: use CloudInitVersion field in the domainmgr
There is a nasty behavior observed on the Vm with a cloud-init
configuration: the field config.UUIDandVersion.Version is
incremented by one each time an application (VM) restarts
from the zedcontrol, which causes cloud-init to be triggered
in the guest and repeat the whole init procedure which is
catastrophic for the guest.

We can't prevent zedcontrol from incrementing the version field,
because this is part of the spec, but what we can do is introduce
another field dedicated for the cloud-init versionning.

This won't work without corresponding fix on the zedcontrol:
in order to keep old domain configs compatible and fix the
nasty behaviour zedcontrol has to update a new cloud-init version
field once with the old version, i.e. something like this
on the zedcontrol side:

    Update(config):
        ...
	if config.CloudInitVersion == 0:
	    // This happens once for old configs where the cloud-init
	    // version field is missing
	    config.CloudInitVersion = config.UUIDandVersion.Version

	config.UUIDandVersion.Version++

On the EVE side we start using a new config.CloudInitVersion
field once it is not zero, which guarantees us that we always
backward and forward compatible.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>