SRIOV VF function PCI is determined by its name
There are cases where for the given VF cloud want to send adapter
assignment before it got back results with PCI address of the device
during the onboard. For that particular case we can determine PCI
address of the given virtual function (VF) using mneumonic in its own
name. VF is named following this rule {VF id}vf{PF IFace} so from that
we can determing Physical Function (PF) to which this VF is belonging to
and VF id. And using that information we can find VF PCI address by
resolving following link

/sys/class/net/{PF IFace}/device/vf{VF id}

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>