pkg/pillar: Fix the maxvcpus option in the xen VMs config files.
The xl.cfg files that are used to configure the Xen guests have the
`maxvcpus` option. The option is used to allow the guest to bring up
more VCPUs than it has online on its start (specified by the `vcpus`
option).

There was a typo in the option string generated by pillar. Fix it.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>