pkg/pillar: Parse the maxcpus controller config option.
The maxcpus VM config option may come from a controller and represents
the maximal amount of VCPUs that a Xen guest can create. The logic for
the corresponding xl.cfg parameter (`maxvcpus`) creation was already
introduced before, hence we need only to parse the parameter and fill
the corresponding field in DomainConfig.

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>