kdump: collect dmesg along with a kdump
Colect plain-text dmesg log along with a kernel dump in order
to simplify kernel debugging: sometimes it can be enough to
take a look into a dmesg once in order to understand a root
cause rather than to download kernel debug symbols, use the
crash tool and dive deeper into the analysis.

And yes, I renamed the 'kcrash-*' to 'kdump-*' in order to
have both files aligned, thus the log looks perfect now:

   >>>>>>>>>>>> Crashed kernel dmesg END <<<<<<<<<<
   kdump collected: /persist/kcrashes/kdump-2022-11-30-13-58-40.dump
   dmesg collected: /persist/kcrashes/dmesg-2022-11-30-13-58-40.log

Promise not to change filenames anymore!

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>