Do not allow switch NI with multiple ports
It is confusing to allow specifying "uplink" or "freeuplink" for the
port for a switch network instance since it will pick an arbitrary port
and there is no check whether the port is up and usable (the probe logic
assumes we have a local IP address to do the probes hence is not and can
not be used on a switch network instance).

While we could block using the labels "uplink" and "freeuplink" that
might cause unneeded errors/failures for configurations where there is
only one uplink aka management port. So instead we check how many ports
are returned when we get the list and reject if > 1.

Signed-off-by: eriknordmark <erik@zededa.com>