nodeagent: log reboot-stack lines tagging with 'dmesg' or 'stack'
'BookReasonKernel' is special and is set by the pkg/kdump once
the kernel dump and dmesg are collected. In the following patch
/persist/reboot-stack will contain the whole dmesg for the
'BootReasonKernel' case, so prepend 'dmesg' string for each line
for easy logs grep.

If kdump is not collected (e.g. ARM arch or xen), then mark a boot
reason with "no kdump" string.

For the pillar backtraces tag each log line with 'stack' prefix.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>