downloader: prepare 'handleSyncOp' for being called in a loop for list of datastores
In the following patches a list of datastores will be used for downloading
the same content tree. This patch prepares the 'handleSyncOp' and moves
datastore and status publishing to the caller, so that 'handleSyncOp'
becomes idempotent, because the main task for this function becomes
to download content or return an error if any, cleaning a garbage
afterwards.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>