pillar: replace all other uses of the 'DatastoreID' with the list
For all log objects manipulation the list of datastores is
concatenated in a single string of uuids.

For the sha resolver we use the first element of the list.
Unfortunately this is a hacky approach and is used only
for the purpose not to break the existing code: we still
depend on the singleton datastore ID field, albeit now
we have a list of datastores.

The main problem is that the singleton datastore ID is
used along the whole downloader stack in different cases
and for now I do not see a simple way how to use the list.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>