wwan (QMI): do not leak client IDs
CID stands for "Client ID". Before using any of the QMI services
(except for the internal CTL service) it is necessary to acquire a CID.

By default, allocation and release of CIDs is done internally by libqmi
via the CTL service. When you run a qmicli command, these operations
are done internally and you do not see them, only the output of the QMI
command you requested is presented.

However, we wrap qmicli call with "timeout" to make sure that if
a command gets stuck (which happens quite often), it will not hang
indefinitely. But if a command timeouts and gets killed, it does not get
the opportunity to release the allocated client ID and it will leak.

The number of available client IDs is quite limited - some services,
like NAS, may provide only 10. After we allocate (and leak) all of them,
we are no longer able to use the service (getting 'ClientIdsExhausted'
error) and a modem reboot is required to fix the issue.

The patch makes the allocation and release of client IDs explicit, i.e.
our code takes care of that, properly handling even the scenario of
a timeouted qmicli call.

Signed-off-by: Milan Lenco <milan@zededa.com>