sriov: Give VFs a different AssignmentGroup than their parent
Currently, VFs get assigned the same AssignmentGroup as their parent.
However, the whole point of VFs is to be able to pass them into
different workloads. By putting them all in the same AssignmentGroup,
they all must be passed through, which defeats the purpose.

Further, VFs should always be in a different IOMMU group anyways, so
this is a safe change.

With this change I'm able to successfully pass in VFs to my VMs running
on EVE.

Signed-off-by: Harsh Modi <hmodi@aviatrix.com>