Avoid spurious download 'cancelled by user'
We've seen one case of this failure and reviewing the code it is clear
that it needs to be more robust against different order on the done and
cancel channels; when the work is done the cancel channel will be
closed. In addition, should the cancel channel be closed it should not
loop reading from it. Also adding logging so we can catch if this
happens and can explain the spurious error.

Signed-off-by: eriknordmark <erik@zededa.com>