wwan: More robust SIM status
Currently, we only look at the first instance of "Application state".
However, the "ready" application can be any of the other applications.
The application that is "ready" is pointed out by the "Primary GW" field
returned in the status.

Therefore, this commit parses the Primary GW field, and extracts the
"Application state" for the slot and application specified in the
Primary GW field

Signed-off-by: Harsh Modi <hmodi@aviatrix.com>