Split netdump publish interval to two config properties
As part of the review process it was requested to split config property
that was previously called "netdump.topic.publish.interval" into two,
one for the publish period before onboarding and another for the publish
period after onboarding.

Previously, the single config property had two different default values
and that would be probably quite confusing for the users.

Signed-off-by: Milan Lenco <milan@zededa.com>