Use NetworkMonitor in zedrouter
NetworkMonitor provides all features needed by zedrouter to watch for
changes in the network stack (used mostly for policy based routing
purposes). This way we can remove some code from the zedrouter and
avoid duplication. Moreover, NetworkMonitor is an interface and there
will be multiple implementations for different network stacks
(currently there is only one for Linux).

This is just a small step in the process of zedrouter refactoring.

Signed-off-by: Milan Lenco <milan@zededa.com>