Retry zedagent cert fetch if it failed
It can fail if there is a network outage when zedagent starts this task.
We run with a short timer when there is a failure, and revert to the
longer timer once we have fetched the initial, and after a trigger or
periodic, succeeded in the fetch.

Signed-off-by: eriknordmark <erik@zededa.com>

Backport to the 8.5 branch

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>