Fix the order of dhcpcd arguments for static DNS config
For some strange reason the order of dhcpcd arguments
"domain_name_servers" and "domain_name" matters. In fact, domain name
should be configured first, otherwise the list of DNS servers is not
applied (or maybe gets reset?). I couldn't find the root cause for this
odd behaviour in the dhcpcd source code or any explanation in the
documentation. I decided to at least report the issue:
https://github.com/NetworkConfiguration/dhcpcd/issues/184

Signed-off-by: Milan Lenco <milan@zededa.com>