handleconfig: move config URL definition into the getLatestConfig()
getLatestConfig() function is by no means a generic function, it is
tightly coupled with the whole logic of how config is parsed, is read
from the file, etc, so we can construct and define the URL inside the
function. This will be especially useful in the next patches, where
config can be read also from the LOC server, so two sources of config
will be supported.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>