Determine controller cert change inside parseControllerCerts
parseControllerCerts already checks if there is any change in newly
fetched controller certificates. We can therefore reuse that code
instead of reloading saved signing certificate.
Plus I'm not sure if we can simply skip parseControllerCerts
when we detect that signing certificate has not changed - what if some
other certificate of a different type has changed? Any cert update
should be published to pubControllerCert.
It is better to do some redundant save of the signing cert (when some
other cert have changed, not the one for signing), then to forget
to publish a cert update.

Signed-off-by: Milan Lenco <milan@zededa.com>

Backport to the 8.5 branch

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>