handleconfig: request device configuration from the LOC
Request configuration from the LOC if attempt to get the configuration
from the controller has failed, but initial configuration was received
either from the controller (onboarded), either successfully read from
the file (/lastconfig).

The LOC config won't be used if it is obsolete, see the corresponding
check in the parseConfig() function.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>