zedagent: create different deferred requests according to the destination
For now we have two destinations for the "info" message: actual
controller (cloud) and local server. By default "info" message
is not periodic, but is sent only on update, which causes a
problem for the stateless LOC implementation (not part of this
patch set).

The idea of a new 'destinationBitset' enum is to have a
controller and a local server be served as usual - by the "info"
messages updates, and only LOC should receive periodic "info"
messages. In order to send periodic "info" messages only to
LOC we need to define destination type (the purpose of this
patch).

This patch defines a type for different destinations and a
function 'queueInfoToDest()' which schedules deferred
requests and runs to a completion for the cloud case and
kicks the timer and offloads the execution for the LOC case.

The LOC deferred request is special: created with the ignore
error flag and runs to a completion in the separate goroutine.

In the following patches type will be used for all Publish*()
calls.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>