vtpm server : use 4 bytes length prefix for protobuf messages
C++ ReadVariant32() is used to read length of the message, but ReadVariant32
just reads the first byte of the wire-formatted protobuf message (varint),
not the length! This is specially problematic when serializing messages
using protobuf binding in Golang (same goes for WriteVariant32).

For example marshaling a EveTPMRequest message in Go, with Command set to
"tpm2_getcap algorithms", will result in the following wire encoding:

'\x0a\x12\x74\x70\x6d\x32\x5f\x63\x72\x65\x61\x74\x65\x70\x72\x69\x6d\x61\x72\x79'

Reading the first byte as base 2, '1010', and dropping the MSB results in
'010', which is 2 in decimal. 2 is wire-type LEN and it is followed by
18 (0x12) the correct size of the message [1]. But C++'s ReadVariant32()
if used for determining the length, reads the message length incorrectly
as 10 (0x0a).

To avoid this issue and not relying on protobuf internals, use standard
4 bytes length prefix as suggested by [2].

[1] https://protobuf.dev/programming-guides/encoding/
[2] https://protobuf.dev/programming-guides/techniques/#streaming

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>