Remove strongSwan-based implementation of Cloud Network Instance
The Cloud Network Instance (L3 network with an IPSec connection
to a remote VPN endpoint) is not covered by tests and has never been
used in production.
Instead, the prevailing opinion is that VPN connectivity (and any other
more advanced networking service) should be provided as a part of
an application deployed on top of EVE.
Even if we decide to keep native IPSec support inside of EVE, it is
easier to first refactor the core of zedrouter and then bring it back
piece by piece. But in that case we should also add some Eden tests
for Cloud Network Instances to have this feature regularly exercised.

Signed-off-by: Milan Lenco <milan@zededa.com>