Network Instance State Collector
The process of collecting state data and metrics for network instances
(IP assignments of app interfaces, interface metrics, application flows)
has been previously scattered all across the zedrouter codebase.
This commit consolidates stata data collecting into a standalone
component defined by an interface nistate.Collector, meaning that the
default Linux-based implementation is replaceable, which opens up
the possibility of adding native support for other network stacks to EVE.
Another important goal of this refactoring was to simplify the existing
implementation and make it easier to understand. Previously, state data
were collected by multiple Go routines and exchanged through some
additional channels and pubsub topics. It was quite difficult to
understand the code as a whole and to see what actually is the
intended behaviour of some of its parts.

Signed-off-by: Milan Lenco <milan@zededa.com>