pillar: hypervisor: Change virtual video display for ARM
The ramfb virtual framebuffer device currently used in QEMU's virt model in
the configuration generated by pillar it's just a very simple frambuffer
driver that must be initialized by the bootloader. Currently VMs are not
detecting the ramfb and therefore VNC doesn't work. According to QEMU's
documentation[1], a good option is to use the virtio-gpu-pci as a display
device.

This commit changes the video device for virt model to virtio-gpu-pci. VMs
are now detecting the virtual device and VNC works.

[1] https://wiki.qemu.org/Documentation/Platforms/ARM#virt_machine_graphics

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>