pkg: kernel: Add patches to fix a bug in EQOS module
This commit adds patches to fix a bug in EQOS module from i.MX8MP platform.
The whole system freezes when the EQOS ethernet interface is added to a
bridge. These patches provide some fixes, improvements and a workaround for
this bug. The driver is now working with bridge mode and can be enabled in
the kernel.

Fix was tested with two devices:

- Phytec phyboard-POLLUX
- Advantech EPC-R3720

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>