yetus: disable hadolint DL3059
this check informs that a consolidation of "RUN"s in
Dockerfiles should be considered, but even the developers
of hadolint themselves admit that there are drawbacks:
https://github.com/hadolint/hadolint/wiki/DL3059

Besides that it disables the use of the cache of docker it
IMO makes the developer write less-readable code with long
concatenations of shell commands for one "RUN".

Another advantage of splitting those lines into several
"RUN"s is that every command runs in it's own shell and
is therefore more independent from other commands
(f.e. environment variables). This makes it more visible
when such dependencies exist.

Docker now also supports squashing the layers after the build
and it could be considered to be used here, but still there is
some effort for development:
https://docs.docker.com/engine/reference/commandline/build/#options

Signed-off-by: Christoph Ostarek <christoph@zededa.com>